Rename SavedGameDataSection2 to SavedGameUnits
