+++ OP2Utility.vcxproj
@@ -131,7 +131,7 @@
     <ClInclude Include="src\Map\CellType.h" />
     <ClInclude Include="src\Map\Map.h" />
     <ClInclude Include="src\Map\MapHeader.h" />
-    <ClInclude Include="src\Map\SavedGameDataSection2.h" />
+    <ClInclude Include="src\Map\SavedGameUnits.h" />
     <ClInclude Include="src\Map\TerrainType.h" />
     <ClInclude Include="src\Map\Tile.h" />
     <ClInclude Include="src\Map\TileGroup.h" />
@@ -160,7 +160,7 @@
     <ClCompile Include="src\Map\MapHeader.cpp" />
     <ClCompile Include="src\Map\MapReader.cpp" />
     <ClCompile Include="src\Map\MapWriter.cpp" />
-    <ClCompile Include="src\Map\SavedGameDataSection2.cpp" />
+    <ClCompile Include="src\Map\SavedGameUnits.cpp" />
     <ClCompile Include="src\ResourceManager.cpp" />
     <ClCompile Include="src\Archive\AdaptiveHuffmanTree.cpp" />
     <ClCompile Include="src\Archive\BitStreamReader.cpp" />
