+++ src/Map/MapReader.cpp
@@ -1,5 +1,5 @@
 #include "Map.h"
-#include "SavedGameDataSection2.h"
+#include "SavedGameUnits.h"
 #include "MapHeader.h"
 #include "../Stream/FileReader.h"
 #include <iostream>
