+++ src/Map/SavedGameUnits.cpp
@@ -1,4 +1,4 @@
-#include "SavedGameDataSection2.h"
+#include "SavedGameUnits.h"
 #include <stdexcept>
 
 void SavedGameUnits::CheckSizeOfUnit() const
