+++ test/Map/SavedGameUnits.test.cpp
@@ -1,4 +1,4 @@
-#include "Map/SavedGameDataSection2.h"
+#include "Map/SavedGameUnits.h"
 #include <gtest/gtest.h>
 #include <stdexcept>
